+++ EssentialFeed/EssentialFeedTests/Feed API/RemoteFeedImageDataLoaderTests.swift
@@ -8,44 +8,6 @@
 import XCTest
 import EssentialFeed
 
-class RemoteFeedImageDataLoader {
-    
-    private let client: HTTPClient
-    
-    init(client: HTTPClient) {
-        self.client = client
-    }
-    
-    public enum Error: Swift.Error {
-        case invalidData
-    }
-    
-    private struct HTTPTaskWrapper: FeedImageDataLoaderTask {
-        let wrapped: HTTPClientTask
-        
-        func cancel() {
-            wrapped.cancel()
-        }
-    }
-    
-    @discardableResult
-    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
-        return HTTPTaskWrapper(wrapped: client.get(from: url) { [weak self] result in
-            guard self != nil else { return }
-            
-            switch result {
-            case let .success((data, response)):
-                if response.statusCode == 200 && !data.isEmpty {
-                    completion(.success(data))
-                } else {
-                    completion(.failure(Error.invalidData))
-                }
-            case let .failure(error): completion(.failure(error))
-            }
-        })
-    }
-}
-
 class RemoteFeedImageDataLoaderTests: XCTestCase {
     
     func test_init_doesNotPerformAnyURLRequest() {
